requirements.txt
aiogram~=3.0.0rc1
tortoise-orm
tortoise-orm[asyncmy]
aerich
gspread-asyncio
environs